Update of the structure of paragraphs
It was decided that the topic "how to get yourself an ORCID should be merged since was composed for just few sentences and also it is an option Learning Objectives. Still related to this topic was decided to emphasize what institutions are already requiring the authors to have an ORCID to reinforce  the importance and impact of this type of ID. As consequence of all this changes, the topic "How to get a PID" had the title updated to emphasize the ID type that we want to advocate for as most relevant at the moment.